+++ cleo_plugins/FileSystemOperations/FileUtils.cpp
@@ -1,4 +1,5 @@
 #include "FileUtils.h"
+#include "CLEO_Utils.h"
 #include <string>
 
 DWORD File::FUNC_fopen = 0;
@@ -84,6 +85,51 @@ bool File::flush(DWORD handle)
 
 DWORD File::open(const char* filename, const char* mode, bool legacy)
 {
+	// validate the mode argument
+	if (!legacy)
+	{
+		static char modeUpdated[12];
+		const std::string allowed = "+abcnrtwxDRST"; // https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/fopen-wfopen?view=msvc-170
+
+		bool valid = false;
+		bool binary = false;
+		bool text = false;
+		auto modeLen = mode != nullptr ? strlen(mode) : 0;
+		if (modeLen > 0 && modeLen < (sizeof(modeUpdated) - 1)) // keep space for extra binary mode char
+		{
+			valid = true;
+
+			for (auto ch : std::string_view(mode))
+			{
+				if (allowed.find(ch) == std::string_view::npos)
+				{
+					valid = false;
+					break; // invalid character
+				}
+
+				if (ch == 'b') binary = true;
+				if (ch == 't') text = true;
+			}
+
+			if (binary && text) valid = false;
+
+			// By default open as binary mode.
+			// Generally text mode is not well documented in C and many file related functions has undefined behavior. For example 'ftell' returns invalid values.
+			if (valid && !binary)
+			{
+				strcpy(modeUpdated, mode);
+				strcat(modeUpdated, "b");
+				mode = modeUpdated;
+			}
+		}
+
+		if (!valid)
+		{
+			LOG_WARNING(0, "Invalid mode argument '%s' while opening file \"%s\" stream!", mode, filename);
+			return 0; // invalid handle
+		}
+	}
+
 	FILE* file = nullptr;
 	if (legacy)
 	{